requirements.txt
streamlit
psycopg[binary]
python-dotenv
anthropic
pandas
plotly
